Refactor consumer and manager into components
