+++ astoria/managers/astdiskd/disk_manager.py
@@ -4,9 +4,9 @@ import logging
 from pathlib import Path
 from typing import List
 
+from astoria.common.components import StateManager
 from astoria.common.disks import DiskInfo, DiskType
 from astoria.common.ipc import DiskManagerMessage
-from astoria.common.manager import StateManager
 
 from .disk_provider import DiskProvider
 from .static import StaticDiskProvider
