+++ astoria/managers/astmetad/metadata_manager.py
@@ -4,13 +4,13 @@ import asyncio
 import logging
 from typing import Dict, List, Optional, Set, Tuple, Type
 
+from astoria.common.components import StateManager
 from astoria.common.disks import DiskInfo, DiskType, DiskUUID
 from astoria.common.ipc import (
     MetadataManagerMessage,
     MetadataSetManagerRequest,
     RequestResponse,
 )
-from astoria.common.manager import StateManager
 from astoria.common.metadata import Metadata
 from astoria.managers.mixins.disk_handler import DiskHandlerMixin
